+++ CHANGELOG.md
@@ -22,7 +22,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Changed
 - Moved permissions functions to [tanjun.permissions][] while keeping deprecated aliases in the now
-  deprecated module [tanjun.utilities][]
+  deprecated module `tanjun.utilities`.
 
 ### Fixed
 - Inherit the global states for `default_member_permissions` and `is_dm_enabled`