+++ CHANGELOG.md
@@ -5,6 +5,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+
+## [2.6.2a1] - 2022-09-05
 ### Added
 - [tanjun.abc.CLient.load_directory][] and [tanjun.abc.Client.load_directory_async][] for loading all
   the modules in a directory.
@@ -652,7 +654,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Removed
 - Removed a lot of impl specific setting and with methods from the abstract interfaces to avoid
 
-[Unreleased]: https://github.com/FasterSpeeding/Tanjun/compare/v2.6.1a1...HEAD
+2.6.2a1
+[Unreleased]: https://github.com/FasterSpeeding/Tanjun/compare/v2.6.2a1...HEAD
+[2.6.2a1]: https://github.com/FasterSpeeding/Tanjun/compare/v2.6.1a1...v2.6.2a1
 [2.6.1a1]: https://github.com/FasterSpeeding/Tanjun/compare/v2.5.4a1...v2.6.1a1
 [2.5.4a1]: https://github.com/FasterSpeeding/Tanjun/compare/v2.5.3a1...v2.5.4a1
 [2.5.3a1]: https://github.com/FasterSpeeding/Tanjun/compare/v2.5.2a1...v2.5.3a1